extract paste backup phrase button into a widget
