+++ lib/routes/initial_walkthrough/mnemonics/enter_mnemonics_page.dart
@@ -1,11 +1,8 @@
 import 'package:breez_translations/breez_translations_locales.dart';
 import 'package:breez_translations/generated/breez_translations.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:misty_breez/routes/routes.dart';
-import 'package:misty_breez/utils/constants/wordlist.dart';
 import 'package:misty_breez/widgets/back_button.dart' as back_button;
-import 'package:misty_breez/widgets/widgets.dart';
 
 class EnterMnemonicsPageArguments {
   final List<String> initialWords;
@@ -73,14 +70,8 @@ class EnterMnemonicsPageState extends State<EnterMnemonicsPage> {
             },
           ),
           actions: <Widget>[
-            IconButton(
-              icon: const Icon(
-                Icons.content_paste_outlined,
-                color: Colors.white,
-                size: 20,
-              ),
-              tooltip: 'Paste Backup Phrase',
-              onPressed: _pasteBackupPhrase,
+            PasteBackupPhraseButton(
+              textEditingControllers: textEditingControllers,
             ),
           ],
           title: Text(
@@ -109,43 +100,5 @@ class EnterMnemonicsPageState extends State<EnterMnemonicsPage> {
         ),
       ),
     );
-  }
-
-  void _pasteBackupPhrase() async {
-    try {
-      final ClipboardData? clipboardData = await Clipboard.getData(Clipboard.kTextPlain);
-      final String? clipboardText = clipboardData?.text?.trim();
-
-      if (clipboardText == null) {
-        throw 'Clipboard is empty.';
-      }
-
-      final List<String> words = _extractWords(clipboardText);
-
-      if (_isValidMnemonic(words)) {
-        _populateTextFields(words);
-      } else {
-        throw 'Clipboard has invalid backup phrase.';
-      }
-    } catch (e) {
-      if (mounted) {
-        showFlushbar(context, message: e.toString());
-      }
-    }
-  }
-
-  List<String> _extractWords(String text) {
-    return text.split(RegExp(r'\s+')).map((String word) => word.toLowerCase().trim()).toList();
-  }
-
-  bool _isValidMnemonic(List<String> words) {
-    return words.length == 12 && words.every((String word) => wordlist.contains(word));
-  }
-
-  void _populateTextFields(List<String> words) {
-    for (int i = 0; i < words.length; i++) {
-      textEditingControllers[i].text = words[i];
-    }
-    FocusManager.instance.primaryFocus?.unfocus();
   }
 }
