+++ lib/routes/initial_walkthrough/mnemonics/widgets/widgets.dart
@@ -1,5 +1,6 @@
 export 'mnemonic_item.dart';
 export 'mnemonic_seed_list.dart';
+export 'paste_backup_phrase_button.dart';
 export 'restore_form.dart';
 export 'restore_form_page.dart';
 export 'verify_form.dart';
